# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Task & QC Management System API
 * OpenAPI spec version: 0.2.0
 */
export interface LoginInput {
  userId: number;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  engineer: "engineer",
  supervisor: "supervisor",
  site_manager: "site_manager",
  technician: "technician",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
}

export interface LoginResult {
  token: string;
  user: User;
}

export interface HealthStatus {
  status: string;
}

export interface Asset {
  id: number;
  name: string;
  model: string;
}

export interface AssetSection {
  id: number;
  assetId: number;
  name: string;
  description?: string;
  order: number;
}

export interface AssetStage {
  id: number;
  sectionId: number;
  name: string;
  stageNumber: number;
  bladeCountMin?: number;
  bladeCountMax?: number;
}

export interface AssetComponent {
  id: number;
  stageId: number;
  name: string;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  draft: "draft",
  assigned: "assigned",
  in_progress: "in_progress",
  paused: "paused",
  submitted: "submitted",
  under_qc: "under_qc",
  approved: "approved",
  rejected: "rejected",
  overdue: "overdue",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string;
  assetId?: number;
  assetName?: string;
  sectionId?: number;
  sectionName?: string;
  stageId?: number;
  stageName?: string;
  stageNumber?: number;
  bladeCountMin?: number;
  bladeCountMax?: number;
  componentId?: number;
  componentName?: string;
  assignedToId?: number;
  assignedToName?: string;
  createdById?: number;
  estimatedHours?: string;
  deadline?: string;
  priority: TaskPriority;
  status: TaskStatus;
  totalMinutes?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface ComponentHistory {
  componentId: number;
  componentName: string;
  stageName?: string;
  sectionName?: string;
  totalTasks: number;
  completedTasks: number;
  avgRepairHours?: number;
  lastMaintenanceDate?: string;
  tasks: Task[];
}

export interface TimeEntry {
  id: number;
  taskId: number;
  userId: number;
  userName?: string;
  startTime: string;
  endTime?: string;
  durationMinutes?: number;
  pauseReason?: string;
  isActive: boolean;
}

export type QcReviewDecision =
  (typeof QcReviewDecision)[keyof typeof QcReviewDecision];

export const QcReviewDecision = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface QcReview {
  id: number;
  taskId: number;
  reviewerId: number;
  reviewerName?: string;
  decision: QcReviewDecision;
  comments?: string;
  createdAt: string;
}

export type TaskDetail = Task & {
  timeEntries?: TimeEntry[];
  qcReviews?: QcReview[];
  activeTimeEntry?: TimeEntry;
};

export type DashboardStatsByStatus = { [key: string]: number };

export type DashboardStatsBySectionItem = {
  sectionName: string;
  count: number;
};

export type DashboardStatsByStageItem = {
  stageName: string;
  count: number;
};

export type DashboardStatsTechnicianPerformanceItem = {
  technicianId: number;
  technicianName: string;
  assignedTasks: number;
  completedTasks: number;
  avgCompletionHours?: number;
};

export interface DashboardStats {
  totalTasks: number;
  byStatus: DashboardStatsByStatus;
  bySection: DashboardStatsBySectionItem[];
  byStage: DashboardStatsByStageItem[];
  technicianPerformance?: DashboardStatsTechnicianPerformanceItem[];
  overdueCount: number;
  approvalRate?: number;
}

export type CreateTaskInputPriority =
  (typeof CreateTaskInputPriority)[keyof typeof CreateTaskInputPriority];

export const CreateTaskInputPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface CreateTaskInput {
  title: string;
  description?: string;
  assetId: number;
  sectionId: number;
  stageId?: number;
  componentId?: number;
  assignedToId?: number;
  estimatedHours?: number;
  deadline?: string;
  priority: CreateTaskInputPriority;
}

export type UpdateTaskStatusInputStatus =
  (typeof UpdateTaskStatusInputStatus)[keyof typeof UpdateTaskStatusInputStatus];

export const UpdateTaskStatusInputStatus = {
  draft: "draft",
  assigned: "assigned",
  in_progress: "in_progress",
  paused: "paused",
  submitted: "submitted",
  under_qc: "under_qc",
  approved: "approved",
  rejected: "rejected",
  overdue: "overdue",
} as const;

export interface UpdateTaskStatusInput {
  status: UpdateTaskStatusInputStatus;
  pauseReason?: string;
  qcComment?: string;
}

export interface StartTimeInput {
  [key: string]: unknown;
}

export interface PauseTimeInput {
  reason: string;
}

export type QcReviewInputDecision =
  (typeof QcReviewInputDecision)[keyof typeof QcReviewInputDecision];

export const QcReviewInputDecision = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface QcReviewInput {
  decision: QcReviewInputDecision;
  comments?: string;
}

export type ListTasksParams = {
  status?: string;
  assignedTo?: number;
  sectionId?: number;
};