# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Maintenance Task & QC Management System API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login by selecting a user (dev mode)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
...
  /assets/{assetId}/sections:
    get:
      operationId: listSections
      tags: [assets]
      summary: List sections for an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetSection"
  /sections/{sectionId}/stages:
    get:
      operationId: listStages
      tags: [assets]
      summary: List stages for a section
      parameters:
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetStage"
  /stages/{stageId}/components:
    get:
      operationId: listComponents
      tags: [assets]
      summary: List components for a stage
      parameters:
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of components
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetComponent"
  /components/{componentId}/history:
    get:
      operationId: getComponentHistory
      tags: [assets]
      summary: Get task history for a component
      parameters:
        - name: componentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Component task history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComponentHistory"
...
  /tasks/{taskId}/qc:
    get:
      operationId: listQcReviews
      tags: [qc]
      summary: List QC reviews for a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QC review history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QcReview"
    post:
      operationId: submitQcReview
      tags: [qc]
      summary: Submit a QC review (approve or reject)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcReviewInput"
      responses:
        "201":
          description: QC review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcReview"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginInput:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    LoginResult:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    AssetSection:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        name:
          type: string
        description:
          type: string
        order:
          type: integer
      required: [id, assetId, name, order]
    AssetStage:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        name:
          type: string
        stageNumber:
          type: integer
        bladeCountMin:
          type: integer
        bladeCountMax:
          type: integer
      required: [id, sectionId, name, stageNumber]
    AssetComponent:
      type: object
      properties:
        id:
          type: integer
        stageId:
          type: integer
        name:
          type: string
      required: [id, stageId, name]
    ComponentHistory:
      type: object
      properties:
        componentId:
          type: integer
        componentName:
          type: string
        stageName:
          type: string
        sectionName:
          type: string
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        avgRepairHours:
          type: number
        lastMaintenanceDate:
          type: string
          format: date-time
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required: [componentId, componentName, totalTasks, completedTasks, tasks]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [engineer, supervisor, site_manager, technician]
      required: [id, name, email, role]
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        assetId:
          type: integer
        assetName:
          type: string
        sectionId:
          type: integer
        sectionName:
          type: string
        stageId:
          type: integer
        stageName:
          type: string
        stageNumber:
          type: integer
        bladeCountMin:
          type: integer
        bladeCountMax:
          type: integer
        componentId:
          type: integer
        componentName:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        createdById:
          type: integer
        estimatedHours:
          type: string
        deadline:
          type: string
          format: date-time
        priority:
          type: string
          enum: [high, medium, low]
        status:
          type: string
          enum: [draft, assigned, in_progress, paused, submitted, under_qc, approved, rejected, overdue]
        totalMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, status, priority, createdAt]
...
    QcReview:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        decision:
          type: string
          enum: [approved, rejected]
        comments:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, taskId, reviewerId, decision, createdAt]
...
    QcReviewInput:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        comments:
          type: string
      required: [decision]