# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Task & QC Management System API
 * OpenAPI spec version: 0.2.0
 */
import type { User } from "./user";

export interface LoginResult {
  token: string;
  user: User;
}