# lib/api-zod/src/generated/types/qcReviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Task & QC Management System API
 * OpenAPI spec version: 0.2.0
 */
import type { QcReviewInputDecision } from "./qcReviewInputDecision";

export interface QcReviewInput {
  decision: QcReviewInputDecision;
  comments?: string;
}